Changed sllist.go
I wanted to be able to get the head and tail of the list
at any time. So the code was refactored to support this.

Before this change was done a list was defined as a node with a
reference to the next node. If i had to track the head and tail, this could
have done via a global variable or inside every node. Which would require
a change to every node if the tail changes.

This new refactored version is a "container" which contains
a reference head and tail of the list.